Migrated main sequence kernels into new kernels module
