+++ diffstar/kernels/sfr_kernels.py
@@ -1,42 +1,15 @@
 """
 """
-from collections import OrderedDict
 from jax import jit as jjit
 from jax import numpy as jnp
 from jax import vmap
-import numpy as np
 from diffmah.individual_halo_assembly import _calc_halo_history
-from ..utils import _sigmoid, _inverse_sigmoid
 from ..utils import jax_np_interp
-from .gas_consumption import _get_lagged_gas
 from .quenching_kernels import quenching_function
-
-_SFR_PARAM_BOUNDS = OrderedDict(
-    lgmcrit=(9.0, 13.5),
-    lgy_at_mcrit=(-3.0, 0.0),
-    indx_lo=(0.0, 5.0),
-    indx_hi=(-5.0, 0.0),
-    tau_dep=(0.0, 20.0),
-)
+from .main_sequence_kernels import _ms_sfr_history_from_mah
 
 TODAY = 13.8
 LGT0 = jnp.log10(TODAY)
-INDX_K = 9.0  # Main sequence efficiency transition speed.
-
-
-def calculate_sigmoid_bounds(param_bounds):
-    bounds_out = OrderedDict()
-
-    for key in param_bounds:
-        _bounds = (
-            float(np.mean(param_bounds[key])),
-            abs(float(4.0 / np.diff(param_bounds[key]))),
-        )
-        bounds_out[key] = _bounds + param_bounds[key]
-    return bounds_out
-
-
-SFR_PARAM_BOUNDS = calculate_sigmoid_bounds(_SFR_PARAM_BOUNDS)
 
 
 @jjit
@@ -248,56 +221,6 @@ calculate_histories_vmap = jjit(
 
 
 @jjit
-def _get_bounded_sfr_params(
-    u_lgmcrit,
-    u_lgy_at_mcrit,
-    u_indx_lo,
-    u_indx_hi,
-    u_tau_dep,
-):
-    lgmcrit = _sigmoid(u_lgmcrit, *SFR_PARAM_BOUNDS["lgmcrit"])
-    lgy_at_mcrit = _sigmoid(u_lgy_at_mcrit, *SFR_PARAM_BOUNDS["lgy_at_mcrit"])
-    indx_lo = _sigmoid(u_indx_lo, *SFR_PARAM_BOUNDS["indx_lo"])
-    indx_hi = _sigmoid(u_indx_hi, *SFR_PARAM_BOUNDS["indx_hi"])
-    tau_dep = _sigmoid(u_tau_dep, *SFR_PARAM_BOUNDS["tau_dep"])
-    bounded_params = (
-        lgmcrit,
-        lgy_at_mcrit,
-        indx_lo,
-        indx_hi,
-        tau_dep,
-    )
-    return bounded_params
-
-
-@jjit
-def _get_unbounded_sfr_params(
-    lgmcrit,
-    lgy_at_mcrit,
-    indx_lo,
-    indx_hi,
-    tau_dep,
-):
-    u_lgmcrit = _inverse_sigmoid(lgmcrit, *SFR_PARAM_BOUNDS["lgmcrit"])
-    u_lgy_at_mcrit = _inverse_sigmoid(lgy_at_mcrit, *SFR_PARAM_BOUNDS["lgy_at_mcrit"])
-    u_indx_lo = _inverse_sigmoid(indx_lo, *SFR_PARAM_BOUNDS["indx_lo"])
-    u_indx_hi = _inverse_sigmoid(indx_hi, *SFR_PARAM_BOUNDS["indx_hi"])
-    u_tau_dep = _inverse_sigmoid(tau_dep, *SFR_PARAM_BOUNDS["tau_dep"])
-    bounded_params = (
-        u_lgmcrit,
-        u_lgy_at_mcrit,
-        u_indx_lo,
-        u_indx_hi,
-        u_tau_dep,
-    )
-    return bounded_params
-
-
-_get_bounded_sfr_params_vmap = jjit(vmap(_get_bounded_sfr_params, (0,) * 5, 0))
-_get_unbounded_sfr_params_vmap = jjit(vmap(_get_unbounded_sfr_params, (0,) * 5, 0))
-
-
-@jjit
 def _integrate_sfr(sfr, dt):
     """Calculate the cumulative stellar mass history."""
     return jnp.cumsum(sfr * dt) * 1e9
@@ -382,52 +305,3 @@ def _sfr_history_from_mah(lgt, dtarr, dmhdt, log_mah, sfr_params, q_params):
     qfrac = quenching_function(lgt, *q_params)
     sfr = qfrac * ms_sfr
     return sfr
-
-
-@jjit
-def _ms_sfr_history_from_mah(lgt, dtarr, dmhdt, log_mah, sfr_params):
-    """Main Sequence formation history of an individual galaxy."""
-
-    bounded_params = _get_bounded_sfr_params(*sfr_params)
-    sfr_ms_params = bounded_params[:4]
-    tau_dep = bounded_params[4]
-    efficiency = _sfr_eff_plaw(log_mah, *sfr_ms_params)
-
-    tau_dep_max = SFR_PARAM_BOUNDS["tau_dep"][3]
-    lagged_mgas = _get_lagged_gas(lgt, dtarr, dmhdt, tau_dep, tau_dep_max)
-
-    ms_sfr = lagged_mgas * efficiency
-    return ms_sfr
-
-
-@jjit
-def _sfr_eff_plaw(lgm, lgmcrit, lgy_at_mcrit, indx_lo, indx_hi):
-    """Instantaneous baryon conversion efficiency of main sequence galaxies
-
-    Main sequence efficiency kernel, epsilon(Mhalo)
-
-    Parameters
-    ----------
-    lgm : ndarray of shape (n_times, )
-        Diffmah halo mass accretion history in units of Msun
-
-    lgmcrit : float
-        Base-10 log of the critical mass
-    lgy_at_mcrit : float
-        Base-10 log of the critical efficiency at critical mass
-
-    indx_lo : float
-        Asymptotic value of the efficiency at low halo masses
-
-    indx_hi : float
-        Asymptotic value of the efficiency at high halo masses
-
-    Returns
-    -------
-    efficiency : ndarray of shape (n_times)
-        Main sequence efficiency value at each snapshot
-
-    """
-    slope = _sigmoid(lgm, lgmcrit, INDX_K, indx_lo, indx_hi)
-    eff = lgy_at_mcrit + slope * (lgm - lgmcrit)
-    return 10**eff
